Use `assertASTEqual` to compare `ast` objects
